pyproject.toml
[tool.poetry]
name = "maxkb"
version = "0.1.0"
description = "智能知识库"
authors = ["shaohuzhang1 <shaohu.zhang@fit2cloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "4.2.15"
djangorestframework = "^3.15.2"
drf-yasg = "1.21.7"
django-filter = "23.2"
langchain = "0.2.3"
langchain_community = "0.2.4"
langchain-huggingface = "^0.0.3"
psycopg2-binary = "2.9.7"
jieba = "^0.42.1"
diskcache = "^5.6.3"
pillow = "^10.2.0"
filetype = "^1.2.0"
torch = {version = "^2.2.1+cpu", source = "pytorch"}
sentence-transformers = "^2.2.2"
openai = "^1.13.3"
tiktoken = "^0.7.0"
qianfan = "^0.3.6.1"
pycryptodome = "^3.19.0"
beautifulsoup4 = "^4.12.2"
html2text = "^2024.2.26"
langchain-openai = "^0.1.8"
django-ipware = "^6.0.4"
django-apscheduler = "^0.6.2"
pymupdf = "1.24.9"
pypdf = "4.3.1"
rapidocr-onnxruntime = "1.3.24"
opencv-python-headless = "^4.10.0"
python-docx = "^1.1.0"
xlwt = "^1.3.0"
dashscope = "^1.17.0"
zhipuai = "^2.0.1"
httpx = "^0.27.0"
httpx-sse = "^0.4.0"
websocket-client = "^1.7.0"
langchain-google-genai = "^1.0.3"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
gunicorn = "^22.0.0"
python-daemon = "3.0.1"
gevent = "^24.2.1"
restrictedpython = "^7.2"
cgroups = "^0.1.0"
wasm-exec = "^0.1.9"

boto3 = "^1.34.151"
langchain-aws = "^0.1.13"
tencentcloud-sdk-python = "^3.0.1205"
xinference-client = "^0.14.0.post1"
psutil = "^6.0.0"
celery = { extras = ["sqlalchemy"], version = "^5.4.0" }
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
celery-once = "^3.0.1"
selenium = "^4.24.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"
